qwen_integration/requirements.txt
# OCR
surya-ocr==0.14.6

# LLM - Qwen3-0.6B specific
transformers>=4.44.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0  # For quantization if needed

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.0.0

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart>=0.0.6
python-magic>=0.4.27

# Utils
pillow>=10.0.0
python-dotenv
typing-extensions
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# For text comparison
python-Levenshtein>=0.21.0